支付

+++ emall/mall-common/src/main/java/edu/hubu/mall/common/utils/PageUtil.java
@@ -15,19 +15,19 @@ public class PageUtil<T> implements Serializable {
     /**
      * 总记录数
      */
-    private int totalCount;
+    private long totalCount;
     /**
      * 每页记录数
      */
-    private int pageSize;
+    private long pageSize;
     /**
      * 总页数
      */
-    private int totalPage;
+    private long totalPage;
     /**
      * 当前页数
      */
-    private int currPage;
+    private long currPage;
     /**
      * 列表数据
      */
@@ -40,7 +40,7 @@ public class PageUtil<T> implements Serializable {
      * @param pageSize    每页记录数
      * @param currPage    当前页数
      */
-    public PageUtil(List<T> list, int totalCount, int pageSize, int currPage) {
+    public PageUtil(List<T> list, long totalCount, long pageSize, long currPage) {
         this.list = list;
         this.totalCount = totalCount;
         this.pageSize = pageSize;
@@ -59,35 +59,35 @@ public class PageUtil<T> implements Serializable {
         this.totalPage = (int)page.getPages();
     }
 
-    public int getTotalCount() {
+    public long getTotalCount() {
         return totalCount;
     }
 
-    public void setTotalCount(int totalCount) {
+    public void setTotalCount(long totalCount) {
         this.totalCount = totalCount;
     }
 
-    public int getPageSize() {
+    public long getPageSize() {
         return pageSize;
     }
 
-    public void setPageSize(int pageSize) {
+    public void setPageSize(long pageSize) {
         this.pageSize = pageSize;
     }
 
-    public int getTotalPage() {
+    public long getTotalPage() {
         return totalPage;
     }
 
-    public void setTotalPage(int totalPage) {
+    public void setTotalPage(long totalPage) {
         this.totalPage = totalPage;
     }
 
-    public int getCurrPage() {
+    public long getCurrPage() {
         return currPage;
     }
 
-    public void setCurrPage(int currPage) {
+    public void setCurrPage(long currPage) {
         this.currPage = currPage;
     }
